+++ CHANGELOG.md
@@ -7,10 +7,13 @@ This project follows [semantic versioning](http://semver.org).
 - [changed] Changed documentation
 - [changed] Updated dependencies
 - [changed] Retrying connections for 120 secs
+- [changed] Resetting own addresses periodically
+- [changed] Using smallvec everywhere
 - [fixed] Fixed corner case with lost init message
 - [fixed] Do not reconnect to timed out pending connections
 - [fixed] Most specific claims beat less specific claims
 - [fixed] Count all invalid protocol traffic
+- [fixed] Fixed compile with musl
 
 ### v2.0.0 (2020-10-30)